+++ probe/endpoint/ebpf.go
@@ -1,23 +1,16 @@
 package endpoint
 
 import (
-	"bufio"
 	"bytes"
 	"encoding/binary"
-	"fmt"
 	"net"
-	"os"
 	"strconv"
-	"strings"
 	"sync"
-	"syscall"
 	"unsafe"
 
 	log "github.com/Sirupsen/logrus"
 	bpflib "github.com/kinvolk/gobpf-elf-loader/bpf"
 )
-
-import "C"
 
 var byteOrder binary.ByteOrder
 
@@ -116,43 +109,6 @@ type EbpfTracker struct {
 	closedConnections []ebpfConnection
 }
 
-func findBpfObjectFile() (string, error) {
-	var buf syscall.Utsname
-	err := syscall.Uname(&buf)
-	if err != nil {
-		return "", err
-	}
-
-	// parse "ID=" in /etc/host-os-release (this is a bind mount of
-	// /etc/os-release on the host)
-	hostDistroFile, err := os.Open("/etc/host-os-release")
-	if err != nil {
-		return "", err
-	}
-	defer hostDistroFile.Close()
-
-	scanner := bufio.NewScanner(hostDistroFile)
-	var distro string
-	for scanner.Scan() {
-		if strings.HasPrefix(scanner.Text(), "ID=") {
-			distro = strings.TrimPrefix(scanner.Text(), "ID=")
-			break
-		}
-	}
-	if err = scanner.Err(); err != nil {
-		return "", err
-	}
-	if distro == "" {
-		return "", fmt.Errorf("distro ID not found")
-	}
-
-	arch := C.GoString((*C.char)(unsafe.Pointer(&buf.Machine[0])))
-	release := C.GoString((*C.char)(unsafe.Pointer(&buf.Release[0])))
-	fileName := fmt.Sprintf("/usr/libexec/scope/ebpf/%s/%s/%s/ebpf.o", distro, arch, release)
-
-	return fileName, nil
-}
-
 func newEbpfTracker(useEbpfConn bool) eventTracker {
 	if !useEbpfConn {
 		return &nilTracker{}
@@ -165,6 +121,9 @@ func newEbpfTracker(useEbpfConn bool) eventTracker {
 	}
 
 	bpfPerfEvent := bpflib.NewBpfPerfEvent(bpfObjectFile)
+	if bpfPerfEvent == nil {
+		return &nilTracker{}
+	}
 	err = bpfPerfEvent.Load()
 	if err != nil {
 		log.Errorf("Error loading BPF program: %v", err)
