+++ probe/endpoint/ebpf_unsupported.go
@@ -1,0 +1,9 @@
+//+build !linux
+
+package endpoint
+
+import "fmt"
+
+func findBpfObjectFile() (string, error) {
+	return "", fmt.Errorf("not supported")
+}
